requirements.txt
kafka==1.3.5
kafka_python==2.0.2
keras==2.6.0
matplotlib==3.6.2
numpy==1.23.5
paho_mqtt==1.6.1
pandas==1.5.2
pykafka==2.8.0
requests==2.27.1
tensorflow==2.6.5
